+++ src/Compilation/ScopeContext.hs
@@ -26,6 +26,7 @@ module Compilation.ScopeContext (
   getProcedureScopes,
 ) where
 
+import Data.List (nub)
 import qualified Data.Map as Map
 import qualified Data.Set as Set
 
@@ -61,7 +62,7 @@ applyProcedureScope ::
   (ScopeContext c -> ScopedFunction c -> ExecutableProcedure c -> a) -> ProcedureScope c -> [a]
 applyProcedureScope f (ProcedureScope ctx fs) = map (uncurry (f ctx)) fs
 
-getProcedureScopes :: (Show c, CollectErrorsM m) =>
+getProcedureScopes :: (Ord c, Show c, CollectErrorsM m) =>
   CategoryMap c -> ExprMap c -> DefinedCategory c -> m [ProcedureScope c]
 getProcedureScopes ta em (DefinedCategory c n pragmas _ _ ms ps fs) = message ??> do
   (_,t) <- getConcreteCategory ta (c,n)
@@ -80,9 +81,13 @@ getProcedureScopes ta em (DefinedCategory c n pragmas _ _ ms ps fs) = message ??
   let cm0 = builtins typeInstance CategoryScope
   let tm0 = builtins typeInstance TypeScope
   let vm0 = builtins typeInstance ValueScope
-  cm2 <- mapMembers readOnly hidden cm
-  tm2 <- mapMembers readOnly hidden $ cm ++ tm'
-  vm2 <- mapMembers readOnly hidden $ cm ++ tm' ++ vm'
+  let immutable = nub $ immutableContext t
+  let readOnly2 = if null immutable
+                     then readOnly
+                     else Map.fromListWith (++) $ Map.toList readOnly ++ zip valueMembers (repeat immutable)
+  cm2 <- mapMembers readOnly2 hidden cm
+  tm2 <- mapMembers readOnly2 hidden $ cm ++ tm'
+  vm2 <- mapMembers readOnly2 hidden $ cm ++ tm' ++ vm'
   mapCompilerM_ checkPragma pragmas
   let cv = Map.union cm0 cm2
   let tv = Map.union tm0 tm2
@@ -104,6 +109,8 @@ getProcedureScopes ta em (DefinedCategory c n pragmas _ _ ms ps fs) = message ??
                                             " does not exist (marked Hidden at " ++
                                             formatFullContext c2 ++ ")") missing
     allMembers = Set.fromList $ map dmName ms
+    valueMembers = map dmName $ filter ((== ValueScope) . dmScope) ms
+    immutableContext t = concat $ map ciContext $ filter isCategoryImmutable (getCategoryPragmas t)
     readOnly = Map.fromListWith (++) $ concat $ map (\m -> zip (mroMembers m) (repeat $ mroContext m)) $ filter isMembersReadOnly pragmas
     hidden   = Map.fromListWith (++) $ concat $ map (\m -> zip (mhMembers m)  (repeat $ mhContext m))  $ filter isMembersHidden   pragmas
     firstM f (x,y) = do